refactor: reorganize card folder structure for clarity
Renamed and relocated card components for better organization:
- Renamed card/root/ to card/home/ for clearer naming
- Moved ProfileIcon to card/portfolio/components/ for better grouping

File moves:
- card/root/* → card/home/
  - AutoCarousel.tsx
  - AutoSlidingBusinessCard.tsx
  - BusinessCard.tsx
- card/portfolio/ProfileIcon.tsx → card/portfolio/components/ProfileIcon.tsx

Updated import paths in:
- src/app/(root)/page.tsx (home card imports)
- src/app/components/project/ProjectPage.tsx (ProfileIcon)
- src/app/(box-layout)/portfolio/Portfolio.tsx (ProfileIcon)
- src/app/(box-layout)/team/Team.tsx (ProfileIcon)
- src/app/components/card/portfolio/components/ProfileIcon.tsx (internal import)

Verified with pnpm lint and pnpm build.
