+++ src/app/components/card/home/AutoCarousel.tsx
